+++ R/Sequence.R
@@ -820,3 +820,42 @@ seqDist <- function(seq1, seq2, dist_mat=getDNAMatrix()) {
 pairwiseDist <- function(seq, dist_mat=getDNAMatrix()) {
     pairwiseDistRcpp(seq, dist_mat)
 }
+
+
+#' Calculate pairwise distances between sequences
+#' 
+#' \code{nonsquareDist} calculates all pairwise distance between a set of sequences and a subset of it.
+#'
+#' @param    seq       character vector containing a DNA sequences.
+#' @param    indx      numeric vector contating the indices (a subset of indices of \code{seq}).
+#' @param    dist_mat  Character distance matrix. Defaults to a Hamming distance 
+#'                     matrix returned by \link{getDNAMatrix}. If gap 
+#'                     characters, \code{c("-", ".")}, are assigned a value of -1 
+#'                     in \code{dist_mat} then contiguous gaps of any run length,
+#'                     which are not present in both sequences, will be counted as a 
+#'                     distance of 1. Meaning, indels of any length will increase
+#'                     the sequence distance by 1. Gap values other than -1 will 
+#'                     return a distance that does not consider indels as a special case.
+#'
+#' @return   A matrix of numerical distance between each entry in \code{seq} and 
+#'           sequences specified by \code{indx} indices. 
+#'           If \code{seq} is a named vector, row and columns names will be added 
+#'           accordingly.
+#' 
+#'           Amino acid distance matrix may be built with \link{getAAMatrix}. 
+#'           Uses \link{seqDist} for calculating distances between pairs.
+#'           See \link{pairwiseEqual} for generating an equivalence matrix.
+#'           
+#' @examples
+#' # Gaps will be treated as Ns with a gap=0 distance matrix
+#' seq <- c(A="ATGGC", B="ATGGG", C="ATGGG", D="AT--C")
+#' pairwiseDist(seq, 
+#'              dist_mat=getDNAMatrix(gap=0))
+#' 
+#' nonsquareDist(seq, indx=c(1,3), 
+#'              dist_mat=getDNAMatrix(gap=0))
+#' 
+#' @export
+nonsquareDist <- function(seq, indx, dist_mat=getDNAMatrix()) {
+    nonsquareDistRcpp(seq, indx, dist_mat)
+}