+++ R/Topology.R
@@ -633,7 +633,7 @@ plotEdgeTest <- function(data, color="black", main_title="Edge Test",
     if (style == "histogram") {
         # Plot edge null distribution
         p1 <- ggplot(perm_sum, aes_string(x="COUNT")) +
-            getBaseTheme() + 
+            baseTheme() + 
             ggtitle(main_title) +
             xlab("Number of edges") +
             ylab("Number of realizations") + 
@@ -644,7 +644,7 @@ plotEdgeTest <- function(data, color="black", main_title="Edge Test",
     } else if (style == "cdf") {    
         # Plot ECDF of edge null distribution
         p1 <- ggplot(perm_sum, aes_string(x="COUNT")) +
-            getBaseTheme() + 
+            baseTheme() + 
             ggtitle(main_title) +
             xlab("Number of edges") +
             ylab("P-value") +
@@ -719,7 +719,7 @@ plotMRCATest <- function(data, color="black", main_title="MRCA Test",
     if (style == "histogram") {
         # Plot MRCA null distribution
         p1 <- ggplot(perm_sum, aes_string(x="COUNT")) +
-            getBaseTheme() + 
+            baseTheme() + 
             ggtitle(main_title) +
             xlab("Number of MRCAs") +
             ylab("Number of realizations") + 
@@ -730,7 +730,7 @@ plotMRCATest <- function(data, color="black", main_title="MRCA Test",
     } else if (style == "cdf") {
         # Plot ECDF of MRCA null distribution
         p1 <- ggplot(perm_sum, aes_string(x="COUNT")) +
-            getBaseTheme() + 
+            baseTheme() + 
             ggtitle(main_title) +
             xlab("Number of MRCAs") +
             ylab("P-value") +
@@ -801,18 +801,13 @@ plotMRCATest <- function(data, color="black", main_title="MRCA Test",
 #' # Define example tree set
 #' graphs <- ExampleTrees[1-10]
 #' 
-#' # Plot violins of outdegree by sample
-#' plotSubtrees(graphs, "SAMPLE", "out", main_title="Node outdegree", 
-#'              style="v")
+#' # Violin plots of node outdegree by sample
+#' plotSubtrees(graphs, "SAMPLE", "out", style="v")
 #'
-#' # Plot violins of subtree by sample
+#' # Violin plots of subtree size by sample
 #' plotSubtrees(graphs, "SAMPLE", "size", style="v")
 #' 
-#' # Plot boxplot of pathlength by isotype
-#' plotSubtrees(graphs, "ISOTYPE", "path", colors=IG_COLORS, 
-#'              legend_title="Isotype", style="b")
-#' 
-#' # Plot boxplot of depth by isotype
+#' # Boxplot of node depth by isotype
 #' plotSubtrees(graphs,  "ISOTYPE", "depth", style="b")
 #' 
 #' @export
@@ -874,7 +869,7 @@ plotSubtrees <- function(graphs, field, stat, root="Germline", exclude=c("Germli
     
     # Make plot object
     p1 <- ggplot(sum_df, aes_string(x=field, y=stat_col)) + 
-        getBaseTheme() + 
+        baseTheme() + 
         ggtitle(main_title) + 
         xlab("") +
         ylab(y_lab) +