+++ inst/markr/build.R
@@ -1,9 +1,0 @@
-library(markr)
-library(alakazam)
-
-# Directories
-pkg_path <- "."
-doc_path <- "./docs"
-
-# Build
-build_mkdocs(pkg_path, doc_path=doc_path, yaml=F)