+++ src/RcppDistance.cpp
@@ -254,3 +254,45 @@ NumericMatrix pairwiseDistRcpp(StringVector seq, NumericMatrix dist_mat) {
     rmat.attr("dimnames") = dimnames;
     return rmat;
 }
+
+
+// nonsquareDist
+// [[Rcpp::export]]
+NumericMatrix nonsquareDistRcpp(StringVector seq, NumericVector indx, NumericMatrix dist_mat)
+{
+    // defien variables
+    int m, n, i, j;
+    int id = 0;
+    std::string col_seq, row_seq;
+    // extract the sizes. Note: This should be satisfied (n<=m)
+    m = indx.size(); //nrow
+    n = seq.size();  //ncolumn
+    // allocate the main matrix and subset matrix
+    NumericMatrix rmat(n,n);
+    std::fill(rmat.begin(), rmat.end(), NA_REAL);
+    rmat.fill_diag(0);
+    NumericMatrix subrmat(m,n);
+    // sort and push indices back by 1 to match c++ indexing
+    std::sort(indx.begin(), indx.end());
+    indx = indx - 1;
+    // begin filling rmat
+    for (i = 0; i < n; i++) {
+        if (!std::binary_search(indx.begin(), indx.end(), i)) continue;
+        col_seq = as<std::string>(seq[i]);     //row sequence 
+        for (j = 0; j < n; j++) {
+            if (!R_IsNA(rmat(i,j))) continue;
+            row_seq = as<std::string>(seq[j]); //col sequence
+            rmat(i,j) = seqDistRcpp(row_seq, col_seq, dist_mat);
+            rmat(j,i) = rmat(i,j);
+        }
+        subrmat(id,_)= rmat(i,_);
+        id++;
+    }
+    // Add row and column names
+    StringVector subSeq = seq[indx];
+    Rcpp::List dimnames = Rcpp::List::create(subSeq.attr("names"),      //rownames
+                                             seq.attr("names"));  //colnames
+    subrmat.attr("dimnames") = dimnames;
+    // return matrix
+    return subrmat;
+}